Unit test fixes: HaxeIntroduceVariableTest.testReplaceAll3 Two primary issues:  The first was a change in the BNF that changed the underlying type of HaxeLocalVarDeclaration to a HaxePsiField.  The second was changing the type of HaxePsiField from a HaxeComponent to a HaxeNamedComponent.
In the first case, the change itself wasn't really a problem; it was that
HaxeComponentType.typeOf() hadn't been changed to reflect the new reality
that it was going to be asked about HAXE_LOCAL_VAR_DECLARATION types.  It
simply returned null for types it didn't know about, thus
HaxeIntroduceHandler.getOccurrences() stopped its loop looking for the
enclosing method element.  The fix was simply to add that type to the list
of classifications.  HAXE_VAR_DECLARATION was also added as future-proofing.

In the second case, HaxePsiField must be a HaxeComponent for
HaxeQualifiedNameProvider to work correctly.  If it is not a HaxeComponent,
a ClassCastException is thrown and refactorings requiring names silently fail.