Add flat icons for nodes and files
This commit adds new flat icons to better match Intellij IDEA's default theme.

The icons are directly based on the official icons for the [nodes] and the [any file type icon]. A 6x6 Haxe logo is added to the bottom right with a 1px transparent outline.

Rasterized versions of the icons are included.

Please note that this commit does not enable the new icons, it only adds the icons to the repository.

Closes HaxeFoundation/intellij-haxe#1020

[nodes]: https://github.com/JetBrains/intellij-community/tree/master/platform/icons/src/nodes
[any file type icon]: https://github.com/JetBrains/intellij-community/blob/master/platform/icons/src/fileTypes/any_type.svg